refactor(interpreter): extract arithmetic evaluation into submodule
Move the recursive-descent arithmetic evaluator and its variable/brace/param
expansion helpers (evaluate_arithmetic*, resolve_arith*, parse_arith*,
parse_arithmetic_impl, parse_base_n, etc.) out of interpreter/mod.rs into
interpreter/arithmetic.rs. Behavior-preserving; moved methods marked
pub(super). ~980 lines moved.
